Use AsyncTaskLoader to perform new show search
Create an AsyncTaskLoader subclass SearchLoader to perform the search,
and display the results in the activity's callback. Replaces using an
AsyncTask.

This is a simple AsyncTaskLoader implementation, which doesn't do
override anything besides forcing a load to start in onStartLoading.
Plenty of examples (in the docs and on stack overflow) override more
functions, which may be required in the future, but this works for now.