Store search results when adding a new show in a singleton
Create a singleton class AddShowSearchResults, which contains a list
of search result objects. The singleton instance replaces the private
list of search results that AddShowSearchActivity used to use. Keeping
the results in a singleton instead of a private member variable will
allow them to be shared with other activities/fragments, i.e. for
displaying details about each search result.

Singletons are not evil in android, apparently.

Synchronisation will probably not be needed around the singleton's
data, as long as it is only read from and written to by the UI thread.