Display a preview of the show when a search result is clicked.
Create AddShowPreviewFragment contained in AddShowPreviewActivity.
These are passed the index of the selected search result via an intent
then fragment argument. They then read the show's information from the
singleton and display its name in the activity's title, and its overview
description in a text view.

If the activity is restarted (eg by leaving it, or rotating the screen),
and if the singleton data has been destroyed (eg due to low memory),
then the activity will finish. This will take the user back to the
search results activity which will restart the search. Not ideal, but
hopefully should not happen very often in real life. A persistent
database may be required instead in the future, however.

Shows can no longer be added, as this now happens instead when a search
result is clicked. This preview activity/fragment needs that
functionality implemented.