Add "add show" item to action bar for preview fragment
Add show to database when it's clicked. This is done synchronously for
the time being, which is okay because only the tvdb id, name, and
overview are added for each show. In the future we'll need to perform
another request to tvdb to gather information about all the episodes,
and maybe some pictures, so this will need to be done asynchronously.