rename tvdb.SearchResult class to tvdb.Show
This class will be extended to contain all fields that a show can have.
It will continue to be used by the search, but will also be used by the
TVDB API calls which return more information about shows. Different API
calls will simply return instances of the Show class with varying levels
of information filled in, as opposed to the original plan of using
different classes for each API call.

Also rename SearchResultParser to SearchShowsParser. Now named after the
function whose result it parses, rather than class it is parsing
instances of from the result.