Add function to get show and episode details from The TVDB.
Add tvdb.Client.getShow(). It takes the show ID as its argument and gets
the full information about the show and all its episodes from The TVDB.
Add the tvdb.Episode class to hold episode information, and add an
episodes list member variable (with a getter and setter) to tvdb.Show.