insert episodes into database when adding a new show
When the add show menu item is selected in AddShowPreviewFragment,
call tvdb.Client.GetShow() to get full information about the show and
its episodes, and insert all that data into the database.

Uses an AsyncTask to perform this. There are issues to do with using
AsyncTasks if they outlive their Activity or Fragment. This should be
fine for the time being, however, especially as it shows no UI while it
is running. In the future we must show some sort of UI and also handle
the lifecycle correctly.