start ShowActivity when show is clicked in ShowListFragment
ShowListFragment propogates its click event to the containing activity
through a OnShowSelectedListener interface. MainActivity then starts
a ShowActivity with the correct show ID.

ShowActivity is a new activity which will display various information
about the show. For the time being it simply displays the show's name
as the activity title. It uses a loader to load the name.