Display list of seasons in ShowActivity
Create a fragment SeasonsListFragment, which displays a list of season
numbers for the specified show. Add this fragment to ShowActivity.

List of seasons is obtained by querying for episodes which match the
show id, and reading their season number column. A MatrixCursor is used
and has a row manually added for each distinct season number found in
the cursor returned from the content provider. This is to work around
the content provider not allowing distinct or group by statements.