add ActionBarSherlock
Add as a git submodule.

Have made a local commit to the submodule though:
Overwrite the support library jar that it bundles in its libs
folder with the same one that we bundle in our libs folder. Also
change its project.properties to build against the same sdk version
as us.

Add a library reference to it in project.properties.