rename pinned shows to starred shows in database and UI
Was concerned about "starring" shows as favourites, because this is
for shows to be sorted first because they are currently being watched,
not because they are all-time favourites. But the word pinned isn't
clear either, and we have star icons available. The word star should be
fine as long as the word favourite isn't also used.

Version 1 databases and new databases will be upgraded or created
properly. Databases which had a pinned column already will break.