replace ActionBarSherlock with ActionBarCompat
ActionBarCompat from the support library v7. Fixes #10.
Bump build and target sdk version to 18 and include newest support
libraries.

Min sdk version should be able to be lowered to 7. But the shows list
item and seasons list item layouts aren't working, so keep at 8.

On old API versions the activity title can't be set properly. This
will be fixed in a later release of the compatibility libraries:
https://code.google.com/p/android/issues/detail?id=58982