don't count unaired episodes in watched progress
it makes it look like the user has episodes available to watch when they
don't. however, we don't want to give the impression that there are no
episodes left of the show at all. so don't count them when setting the
progress bar, but if there are upcoming episodes then say so in the
progress bar's text.

e.g. "Watched 6 of 9 episodes (+ 3 upcoming)"

Episodes with a date in the future or no date at all are being treated
as upcoming, unless they are specials.

fixes #12