set activity titles to empty string initially
for activities which load their title at run time, set the title to an
empty string in the manifest. for the brief time before the correct
title has loaded it's better to show nothing than the default title
"Episodes". (it looks a bit rubbish showing "Episodes" for a split
second then changing it.)