display banner images in shows list fragment
use universal image loader to load banners from thetvdb and display them
instead of the text of the show title.

if the banner is missing or while it is being downloaded continue to
display the show title as text. display a completely transparent image
that has the same dimensions as the banners on top, so that the list
item is the correct size.

use universal image loader's default memory and disk caches.