add auto-refresh settings
Add a new activity SettingsActivity which simply contains an instance of
another new fragment SettingsFragment. Add a menu item to the main
activity to open the settings activity. Slight code cleanup in
MainActivity while we're there.

Add the following settings:
 * Automatically refresh shows - a checkbox that enables or disables
   auto-refresh. Defaults to checked.

 * Auto-refresh frequency - selects the frequency at which the shows are
   automatically refreshed. Can be set to the following options:
    * Daily
    * Twice weekly
    * Weekly
    * Fortnightly
   Defaults to weekly.
   The constants associated with these strings are periods not
   frequencies, so the variables are named as such. The user-facing
   strings refer to them as frequencies.
   Only enabled if "automatically refresh shows" is checked.

This commit only adds these settings, it doesn't implement the
functionality.

Bump minSdkVersion to 11, for PreferenceFragment, as it unfortunately
doesn't exist in the support library.