schedule periodic alarm to auto refresh shows
Create a singelton class AutoRefreshHelper. Make it read and listen to
changes in the settings regarding auto-refresh. When auto-refresh is
enabled schedule a service to run after roughly a certain time, and when
auto refresh is disabled cancel that service. Use AlarmManager to do
this. Use the value of the auto-refresh frequency setting, along with
the time of the last auto-refresh, to determine what time to run the
service at.

The service itself is responsible for refreshing all of the shows,
storing the time at which it last did so, and asking the helper to
schedule it again.

Using the AlarmManager was chosen over using a SyncAdapter for a few
reasons:
 * We do not use accounts, and adding a dummy one to the system--like
   the SyncAdapter approach required--looked very strange.
 * I believe it will be easier to add an "only auto-refresh over Wi-Fi"
   option in the future using this approach.
 * I don't think a SyncAdapter's periodic sync works correctly when it
   is toggled on/off or the phone reboots. So part of the chosen
   approach (storing the time of last refresh) would have had to be
   implemented anyway.