wait until network is available for auto-refresh
If an auto-refresh is attempted but a network connection is not
available then a broadcast receiver is enabled to listen for the network
state change. When the network becomes available again the receiver is
disabled (to conserve battery) and the auto-refresh is rescheduled. This
auto-refresh should now run straight away as it will be overdue from the
previous attempt.