implement restore library from backup
Add a function restore() to BackupRestoreHelper. It copies a backed-up
database to the location of the application's actual database. It then
makes the content provider reopen the database and notify everything
that the data has changed.

Add an item to the main activity's menu which calls this
function. Currently the location of the backup database is hardcoded.