allow user to pick which backup to restore
When the user selects restore from the main menu a dialog is opened
listing all the files in the app's backup folder on the phone's external
storage. If one of these backups is selected then the app data is
restored from it.

If the backup folder is empty then a message saying so is displayed
instead of a list of backups.

There is no protection against selecting any random file that isn't a
previously made backup. Doing so will probably not cause an error but
obviously will not result in any valid data being restored: effectively
the data will be cleared.