Add title strip to ShowActivity
Add a text view with the primary colour as its background colour
directly above the tab strip and below the header image. The tab strip
also has the primary colour as its background so they form a single
area. Display the title of the current show in this text view in fairly
large white text.

Remove the title from the toolbar as it is no longer required. Make the
tabs no longer expand to the entire width as it now looks better when
they don't.