Add build flavours with different applicationIds
Since somebody else uploaded this application to the play store without
changing the applicationId, it needs to be changed for me to release it
on the play store. I do not want to completely change it, however, as
that would be disruptive to F-Droid and existing users.

Add a gradle productFlavor called "playstore" that builds with the
applicationId "com.vindustries.epsiodes". And one called "fdroid" that
builds with the old id "org.jamienicol.episodes". Ensure that the
content provider uses the correct applicationId, meaning both versions
can be installed at the same time.