show/hide episode details fragment on cursor load
show and hide not the fragment's view, but its direct child which holds
everything else. this means that when the next episode fragment is the
current page in show activity you can still scroll it up and down (to
collapse the header image). This also ensures that the checkbox is
invisible before the data is loaded, without treating the checkbox
differently to the other items. Do this for both NextEpisodeFragment and
EpisodeDetailsFragment.