Use language when adding and refreshing shows
Add language parameter to tvdb.Client.getShow() and send it to tvdb. If
null, fall back to "en", so as to not break anything.

When adding a new show, pass the language along with tvdb id and name in
the intent to AddShowService, which will pass it getShow().

When refreshing a show already in the database, first query the database
for the language as we currently do for the tvdb id, then pass it to getShow().