tvdb: Parse language in GetShowParser
Not doing this meant that when adding or refreshing a show, the returned
language was null. This meant that we were always inserting null into
the database as the show's language. That lead to the show's data
reverting to english (or disappearing if the show has no english data on
tvdb) upon first refresh.